+++ mypy-strict.ini
@@ -40,7 +40,6 @@ files =
     .github,
     benchmarks/instruction_counts,
     tools,
-    torch/testing/_internal/framework_utils.py,
     torch/utils/_pytree.py,
     torch/utils/benchmark/utils/common.py,
     torch/utils/benchmark/utils/timer.py,
