+++ mypy.ini
@@ -38,8 +38,8 @@ files =
     test/test_type_hints.py,
     test/test_type_info.py,
     test/test_utils.py,
-    tools/clang_format_utils.py,
-    tools/clang_tidy.py,
+    tools/linter/clang_format_utils.py,
+    tools/linter/clang_tidy.py,
     tools/generate_torch_version.py,
     tools/render_junit.py,
     tools/stats_utils
@@ -109,7 +109,7 @@ warn_unused_ignores = False
 [mypy-tools.generate_torch_version]
 warn_unused_ignores = False
 
-[mypy-tools.stats_utils.s3_stat_parser]
+[mypy-tools.stats.s3_stat_parser]
 warn_unused_ignores = False
 
 #
