+++ tools/README.md
@@ -37,11 +37,11 @@ Build system pieces:
 
 Developer tools which you might find useful:
 
-* [clang_tidy.py](clang_tidy.py) - Script for running clang-tidy
+* [linter/clang_tidy.py](linter/clang_tidy.py) - Script for running clang-tidy
   on lines of your script which you changed.
 * [extract_scripts.py](extract_scripts.py) - Extract scripts from
   `.github/workflows/*.yml` into a specified dir, on which linters such as
-  [run_shellcheck.sh](run_shellcheck.sh) can be run. Assumes that every `run`
+  [linter/run_shellcheck.sh](linter/run_shellcheck.sh) can be run. Assumes that every `run`
   script has `shell: bash` unless a different shell is explicitly listed on that
   specific step (so `defaults` doesn't currently work), but also has some rules
   for other situations such as [actions/github-script][]. Exits with nonzero
@@ -53,17 +53,17 @@ Developer tools which you might find useful:
   can conveniently run diffs on them when working on code-generation.
   (See also [generated_dirs.txt](generated_dirs.txt) which
   specifies the list of directories with generated files.)
-* [mypy_wrapper.py](mypy_wrapper.py) - Run `mypy` on a single file using the
+* [linter/mypy_wrapper.py](linter/mypy_wrapper.py) - Run `mypy` on a single file using the
   appropriate subset of our `mypy*.ini` configs.
-* [run_shellcheck.sh](run_shellcheck.sh) - Find `*.sh` files (recursively) in
+* [linter/run_shellcheck.sh](linter/run_shellcheck.sh) - Find `*.sh` files (recursively) in
   the directories specified as arguments, and run [ShellCheck][] on all of them.
-* [test_history.py](test_history.py) - Query S3 to display history of a single
+* [stats/test_history.py](stats/test_history.py) - Query S3 to display history of a single
   test across multiple jobs over time.
-* [trailing_newlines.py](trailing_newlines.py) - Take names of UTF-8 files from
+* [linter/trailing_newlines.py](linter/trailing_newlines.py) - Take names of UTF-8 files from
   stdin, print names of nonempty files whose contents don't end in exactly one
   trailing newline, exit with status 1 if no output printed or 0 if some
   filenames were printed.
-* [translate_annotations.py](translate_annotations.py) - Read [Flake8][] or
+* [linter/translate_annotations.py](linter/translate_annotations.py) - Read [Flake8][] or
   [clang-tidy][] warnings (according to a `--regex`) from a `--file`, convert to
   the JSON format accepted by [pytorch/add-annotations-github-action], and
   translate line numbers from `HEAD` back in time to the given `--commit` by
