+++ tools/linter/clang_format_utils.py
