+++ tools/linter/git-clang-format
