+++ tools/linter/trailing_newlines.py
