+++ tools/stats/test_history.py
@@ -7,8 +7,8 @@ from datetime import datetime, timezone
 from signal import SIG_DFL, SIGPIPE, signal
 from typing import Dict, Iterator, List, Optional, Set, Tuple
 
-from tools.stats_utils.s3_stat_parser import (Report, get_cases,
-                                              get_test_stats_summaries)
+from tools.stats.s3_stat_parser import (Report, get_cases,
+                                        get_test_stats_summaries)
 
 
 def get_git_commit_history(
@@ -193,7 +193,7 @@ In multiline mode, each line next includes the name of a CircleCI job,
 followed by the time of the specified test in that job at that commit.
 Example:
 
-    $ tools/test_history.py --mode=multiline --ref=594a66 --sha-length=8 --test=test_set_dir \
+    $ tools/stats/test_history.py --mode=multiline --ref=594a66 --sha-length=8 --test=test_set_dir \
       --job pytorch_linux_xenial_py3_6_gcc5_4_test --job pytorch_linux_xenial_py3_6_gcc7_test
     2021-02-10 11:13:34Z 594a66d7 pytorch_linux_xenial_py3_6_gcc5_4_test 0.36s
     2021-02-10 11:13:34Z 594a66d7 pytorch_linux_xenial_py3_6_gcc7_test 0.573s errored
@@ -211,7 +211,7 @@ Example:
 
 Another multiline example, this time with the --all flag:
 
-    $ tools/test_history.py --mode=multiline --all --ref=321b9 --delta=12 --sha-length=8 \
+    $ tools/stats/test_history.py --mode=multiline --all --ref=321b9 --delta=12 --sha-length=8 \
       --test=test_qr_square_many_batched_complex_cuda
     2021-01-07 10:04:56Z 321b9883 pytorch_linux_xenial_cuda10_2_cudnn7_py3_gcc7_test2 424.284s
     2021-01-07 10:04:56Z 321b9883 pytorch_linux_xenial_cuda10_2_cudnn7_py3_slow_test 0.006s skipped
@@ -226,7 +226,7 @@ In columns mode, the name of the job isn't printed, but the order of the
 columns is guaranteed to match the order of the jobs passed on the
 command line. Example:
 
-    $ tools/test_history.py --mode=columns --ref=3cf783 --sha-length=8 --test=test_set_dir \
+    $ tools/stats/test_history.py --mode=columns --ref=3cf783 --sha-length=8 --test=test_set_dir \
       --job pytorch_linux_xenial_py3_6_gcc5_4_test --job pytorch_linux_xenial_py3_6_gcc7_test
     2021-02-10 12:18:50Z 3cf78395    0.644s    0.312s
     2021-02-10 11:13:34Z 594a66d7    0.360s  errored
