+++ tools/test/test_stats.py
@@ -2,11 +2,11 @@
 import unittest
 from typing import Dict, List
 
-from tools import print_test_stats
-from tools.stats_utils.s3_stat_parser import (Commit, Report, ReportMetaMeta,
-                                              Status, Version1Case,
-                                              Version1Report, Version2Case,
-                                              Version2Report)
+from tools.stats import print_test_stats
+from tools.stats.s3_stat_parser import (Commit, Report, ReportMetaMeta,
+                                        Status, Version1Case,
+                                        Version1Report, Version2Case,
+                                        Version2Report)
 
 
 def fakehash(char: str) -> str:
